updated package structure, readme and put method per conversation with Scott this afternoon
