+++ kvcache/kvcache_test.go
@@ -1,14 +1,12 @@
-package main
+package kvcache
 
 import (
 
 	"testing"
 )
 
-
 func TestPut(t *testing.T) {
 	cache := &simpleKeyValueCache{data: map[string]string{}}
-
 
 	cache.Put("name is key","value is Troy Dai")
